# Roca Maria/SparqlEndpoint/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/triples/store": {
            "post": {
                "tags": [
                    "Triples"
                ],
                "summary": "Store job-related triples in the knowledge base",
                "description": "Stores RDF triples for jobs with batch processing and error logging",
                "operationId": "326576dd2b2eee4dd0772f95c32f103f",
                "requestBody": {
                    "description": "Job data for triple generation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "required": [
                                        "jobTitle",
                                        "company",
                                        "location"
                                    ],
                                    "properties": {
                                        "date": {
                                            "type": "string",
                                            "example": "27 May 2024"
                                        },
                                        "jobTitle": {
                                            "type": "string",
                                            "example": "Senior Developer"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Tech Corp"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "New York, USA"
                                        },
                                        "jobDescription": {
                                            "type": "string",
                                            "example": "What you need to know: - .NET Core, ASP.NET MVC, Web API - C# - HTML and CSS"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "example": [
                                {
                                    "date": "12 December 2024",
                                    "jobTitle": "Backend Engineer",
                                    "company": "StartupX",
                                    "location": "Remote",
                                    "jobDescription": "Experience with Node.js, Express, MongoDB"
                                },
                                {
                                    "date": "15 November 2024",
                                    "jobTitle": "Data Scientist",
                                    "company": "AI Labs",
                                    "location": "London, UK",
                                    "jobDescription": "Python, TensorFlow, Keras"
                                }
                            ]
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Triples stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "inserted_triples": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Error processing job 3: Invalid data format"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No data provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "jobTitle": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The jobTitle field is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}